Refactor getting missing transactions
Iterate over each budget's list of transactions separately instead of
using the longer list's length as the iteration count. This saves
repeatedly checking for IndexErrors while iterating and allows for
independent current positions, for example when one budget is missing
transactions the current position of the other budget will need to be
bumped as many times as transactions are missing.