Fix transaction pair bug
Inflow vs outflow wasn't being taken into account. When comparing
transactions from one budget to the other, the amount has to be
inversed. For an inflow, or positive amount, transaction in one budget,
the matching pair transaction in the other budget is an outflow, or
negative amount, one.

The "Borrow for nachos" inflow transaction from other_budget was being
returned as the missing transaction from this_budget when, actually,
this_budget already had the pair "Loan for nachos" outflow transaction.