Fix same-amount missing transaction detection
superset can reach the end before subset transactions if subset has a
transaction not present in superset. In that case, the transactions from
subset not present in superset need to be subtracted from the subset
count when checking the amount of detected missing transactions.

Transactions from subset aren't added to superset because the detection
is uni-directional. Those transactions will be dealt with when the
method is executed with the budgets swapped.