Fix bug when advancing past missing transactions
If the transaction that triggered a detection of missing transactions
isn't the first transaction in the group of transactions for a given
amount, using len(missing_transactions) to advance results in skipping
transactions. E.g.

([-2, -2, -2, -1.8], [2, 1.8])
       ^

The second -2 transaction would detect there are two transactions
missing. In some cases, there could be three or more missing, and the
next transaction would be skipped past -1.8 to another quantity. Then
transactions of amount [-1.8, 1.8] would be tallied, and they might
match, throwing a YnabBudgetComparerValueError.

If missing transactions