Add localbudgets route
Wire up the shell.html view, loaded by index.html to bootstrap the app.
Consequently removed the route config for the layout feature, as it is
not needed.
Add a top menu that dynamically creates entries based on the routes that
exist, according to the loaded modules. Inspired by John Papa's
"modular" demo app.