Fix default view not loading on empty route
Rules added to $urlRouterProvider via when() or otherwise() do not have
any effect unless $state is injected into app.core for some reason.

Change the rule from otherwise() to when(), specifying the empty route
''. This is because the intention is to only redirect the home route to
local budgets, not catch all invalid routes.