Simplify server instantiation
The object hierarchy to run the Flask app was convoluted, with the
Flask object being instantiated in __init__.py and then imported in
the api module. The app module then imported the api module to add
routes for the front end.

This refactor moves the Flask object instantiation to the app module,
and the api module imports it to add API routes.