Update Heroku deploy configuration
Due to the change in wiring up the server, both the api and the app
modules need to be imported before running the Flask server, as they
add functionality to it (API routes and view routes).

Gunicorn only takes a module name and one of its variables as the
argument, so a module is needed that does the needed imports.

Gunicorn imports the module specified rather than executing it, so
__name__ will never be '__main__'. Taking advantage of that, run.py can
be used for both local debug and Heroku deployment (where debug is set
to False).