Add argument validation
Use Flask RESTful reqparse for local budget comparison argument
validation, adding a custom non_empty_string type for the target
categories.

Create a werkzeug custom converter for the dropbox budgets API call, as
the argument is in the URL slug. This enforces either 'this' or 'other'
in the API call URL, returning 404 if another value is sent.